add slice events shim
