+++ packages/job-components/src/operations/index.ts
@@ -2,8 +2,8 @@ import BatchProcessor from './batch-processor';
 import ConvictSchema from './convict-schema';
 import DataEntity from './data-entity';
 import Fetcher from './fetcher';
-import legacyProcessorShim from './shims/legacy-processor';
-import legacyReaderShim from './shims/legacy-reader';
+import legacyProcessorShim from './shims/legacy-processor-shim';
+import legacyReaderShim from './shims/legacy-reader-shim';
 import OperationAPI, { OpAPI, OpAPIInstance, OpAPIFn } from './operation-api';
 import ParallelSlicer, { SlicerFn } from './parallel-slicer';
 import Processor from './processor';
